src/gateway_app/requirements.txt
fastapi
uvicorn
aio-pika
websockets
attrs
sqlalchemy
pydantic-settings
psycopg2-binary
asyncpg
greenlet
celery
redis